Set the new global backend manager.
localization_backend_manager::global() returns a copy of
localization_backend_manager, as opposed to a pointer or reference.
Therefore, when we call localization_backend_manager::select(), we are
mofifying the local copy, and not the global backend manager. Therefore
we need to set the modified copy as the new global backend manager.